Pigeon House
Added the pigeon's house, and added a script to turn the interior on and off, since the model is see through on the inside. Other interior items can be placed in there. Re-structured the hierarchy of the main scene as well.

Silohettes still need to be hidden through the walls in this scene